Move share related functions into a proper file
